pyproject.toml
[tool.poetry]
name = "image-enhancement-srgan"
version = "0.1.0"
description = ""
authors = ["DarrionG <darrionguest@gmail.com>"]
readme = "README.md"
packages = [{include = "image_enhancement_srgan"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
